pyproject.toml
[project]
name = "django-ninja-jsonapi"
version = "0.1.0"
description = "JSON:API toolkit for Django Ninja"
readme = "README.md"
authors = [
    { name = "Ignace Maes", email = "10243652+IgnaceMaes@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "django-ninja>=1.0",
    "orjson>=3.10.0",
    "pydantic>=2.6.0",
]

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
    "ruff>=0.15.2",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = [".", "src"]
django_find_project = false
